+++ lerna.json
@@ -1,5 +1,4 @@
 {
-  "npmClient": "yarn",
   "packages": [
     "packages/*"
   ],
