+++ tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "composite": true
+  },
+  "files": [],
+  "include": [],
+  "references": [
+    { "path": "./packages/unmock" },
+    { "path": "./packages/unmock-jsdom" },
+    { "path": "./packages/unmock-node" }
+  ]
+}
