+++ tsconfig.test.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "composite": true
+  },
+  "files": [],
+  "include": [],
+  "references": [
+    { "path": "./packages/unmock/test" },
+    { "path": "./packages/unmock-jsdom/test" },
+    { "path": "./packages/unmock-node/test" }
+  ]
+}
